requirements.txt
# Core Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0

# LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.16
langchain-anthropic>=0.1.1

# Vector Store
faiss-cpu>=1.7.4
# For cloud: pinecone-client>=3.0.0

# Embeddings
sentence-transformers>=2.3.0

# Document Processing
pypdf>=4.0.0
python-multipart>=0.0.6
beautifulsoup4>=4.12.0
markdown>=3.5.0

# LLM APIs
anthropic>=0.18.0  # Optional - only if using Claude
requests>=2.31.0  # For Hugging Face free API

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
tiktoken>=0.5.0

# Session Management (Optional)
redis>=5.0.0

# Development
pytest>=7.4.0
black>=24.1.0
